ci: update the runner images: ubuntu 22.04 -> 24.04
This should be a very low risk change, since the only CI job that directly uses
the base image is .github/workflows/verify-version-consistency.yml, which only
depends on bash.

All the other workflows run in containers and should be unaffected; we will
update the base image in the next commit.

Installed software on the two platforms:
    - https://github.com/actions/runner-images/blob/main/images/ubuntu/Ubuntu2204-Readme.md
    - https://github.com/actions/runner-images/blob/main/images/ubuntu/Ubuntu2404-Readme.md