ci: build for Windows, and show that Windows builds are currently not supported
This commit adds a `tests-for-windows` workflow that builds a win64 version of
the library with both autotools and CMake, and then attempts to run the frost
example and the functional tests via wine.

In its current state, the builds fail because getrandom() does not exist under
Windows.

The new workflow verifies that the Windows builds are failing, in order to offer
a way of flipping the check when the Windows binaries will be working.