frost: release v0.3.2-frost-2
The version that need to go into itcoin-v25 is v0.3.2-frost-1, and is not based
on the final version of secp256k1 (because bitcoin v25 did so).

This proposed version (v0.3.2-frost-2), instead, incorporates the final
secp256k1 v0.3.2, and includes our recent changes: simplifications in the math
performed in the frost module, multiplatform support and better CI checks.

These changes originated from the upgrading effort to v0.4.0 (#35), that we
decided to apply to 0.3.2 before upgrading the secp256k1 version.