RFC9591: Update compute_challenge()
This commit updates compute_challenge() to follow the specification in RFC-9591 (FROST). The challenge is a hash functino computed on some points. These points are serialized as 33-bytes.

Since the current implementation of FROST support tweaking point (as specified in BIP-340), we performed some tweaking before computing the challenge to avoid issues.