build: move ec point tweak into macros, add --enable-module-frost-bip340-mode and SECP256K1_ENABLE_MODULE_FROST_BIP340_MODE
Introduce a BIP-340 compatibility mode for FROST. This commit creates the
compilation flags for autotools and CMake. The flags control whether the macro
ENABLE_MODULE_FROST_BIP340_MODE is set, which in turn will control if the
library is compiled in BIP340 or RFC9591 mode (the support for RFC9591 will be
finished in the following commits).

For now the default keeps the current behavior, which is BIP340 compatibility.

This commit moves the ec point tweak logic into macros controlled by
ENABLE_MODULE_FROST_BIP340_MODE.