tests: evaluate FROST against RFC9591 test vectors when compiled in standard mode
When the library is compiled in RFC9591-compliant mode (i.e., when
ENABLE_MODULE_FROST_BIP340_MODE is not defined), additional tests are executed,
in order to verify that that the provided FROST implementation complies with the
standard proposed in RFC-9591.

These tests are not executed when the library is compiled in BIP-340 mode.