frost: invert parameter order for compute_hash_h2()
The standard parameter ordering for this library is to put output at the
beginning of the parameter list.

Changed the formal parameter name from "hash_value" to out32 in preparation for
the integration of hash to curve.

The function is currently not invoked in the code base.

No functional changes.